FIX: maths_functions

+++ Start/lib/src/maths/my_factorial.c
@@ -9,13 +9,13 @@
 
 /**
  * @ingroup maths
- * @brief Return the factorial of a double number.
+ * @brief Return the factorial of a number.
  * @param nb The number.
  * @return The factorial of number.
  * @note COmplexity: log(nb)
  * @note Part of UtilsLib by Victor Defauchy.
  */
-double my_factorial(double nb)
+int my_factorial(int nb)
 {
     return (nb == 0) ? 1 : nb * my_factorial(nb - 1);
 }
